renamed runtime test junit package
