+++ monticore-grammar/src/test/java/de/monticore/types3/AbstractTypeVisitorTest.java
@@ -11,8 +11,8 @@ import de.monticore.expressions.lambdaexpressions._ast.ASTLambdaExpression;
 import de.monticore.expressions.lambdaexpressions._symboltable.LambdaExpressionsSTCompleteTypes2;
 import de.monticore.ocl.oclexpressions.symboltable.OCLExpressionsSymbolTableCompleter;
 import de.monticore.ocl.setexpressions.symboltable.SetExpressionsSymbolTableCompleter;
-import de.monticore.runtime.junit.jupyter.AbstractMCTest;
-import de.monticore.runtime.junit.jupyter.MCAssertions;
+import de.monticore.runtime.junit.AbstractMCTest;
+import de.monticore.runtime.junit.MCAssertions;
 import de.monticore.symbols.basicsymbols.BasicSymbolsMill;
 import de.monticore.symbols.basicsymbols._symboltable.FunctionSymbol;
 import de.monticore.symbols.basicsymbols._symboltable.IBasicSymbolsScope;
@@ -43,7 +43,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import static de.monticore.runtime.junit.jupyter.MCAssertions.assertNoFindings;
+import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
 import static de.monticore.types3.util.DefsTypesForTests._booleanSymType;
 import static de.monticore.types3.util.DefsTypesForTests._boxedListSymType;
 import static de.monticore.types3.util.DefsTypesForTests._csStudentSymType;
