+++ monticore-grammar/src/test/java/de/monticore/types3/JavaClassExpressionsTypeVisitorTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.io.IOException;
 import java.util.stream.Stream;
 
-import static de.monticore.runtime.junit.jupyter.MCAssertions.assertNoFindings;
+import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
 
 public class JavaClassExpressionsTypeVisitorTest
     extends AbstractTypeVisitorTest {
