+++ monticore-grammar/src/test/java/de/monticore/types3/ResolveWithinTypeTest.java
@@ -29,7 +29,7 @@ import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import java.util.List;
 
-import static de.monticore.runtime.junit.jupyter.MCAssertions.assertNoFindings;
+import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
 import static de.monticore.types3.util.DefsTypesForTests._booleanSymType;
 import static de.monticore.types3.util.DefsTypesForTests.field;
 import static de.monticore.types3.util.DefsTypesForTests.inScope;
