+++ monticore-grammar/src/test/java/de/monticore/types3/SIUnitTypes4MathTypeVisitorTest.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-import static de.monticore.runtime.junit.jupyter.MCAssertions.assertNoFindings;
+import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
 import static de.monticore.types3.util.SIUnitIteratorForTests.get2UnitsGroup;
 import static de.monticore.types3.util.SIUnitIteratorForTests.getPrefixedUnits;
 import static de.monticore.types3.util.SIUnitIteratorForTests.getUnits;
