+++ monticore-runtime-tests/src/main/java/de/monticore/runtime/junit/AbstractMCTest.java
@@ -1,5 +1,5 @@
 /* (c) https://github.com/MontiCore/monticore */
-package de.monticore.runtime.junit.jupyter;
+package de.monticore.runtime.junit;
 
 import de.se_rwth.commons.logging.Log;
 import de.se_rwth.commons.logging.LogStub;
